+++ src/main/java/org/thema/graphab/metric/local/IFLocalMetric.java
@@ -17,72 +17,49 @@
  */
 
 
-package org.thema.graphab.metric.global;
+package org.thema.graphab.metric.local;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
+import org.geotools.graph.structure.Graphable;
 import org.geotools.graph.structure.Node;
 import org.thema.graphab.Project;
 import org.thema.graphab.graph.GraphGenerator;
+import org.thema.graphab.graph.GraphPathFinder;
 import org.thema.graphab.links.Linkset;
 import org.thema.graphab.metric.AlphaParamMetric;
-import org.thema.graphab.metric.Circuit;
 import org.thema.graphab.metric.ParamPanel;
-import org.thema.graphab.metric.PreCalcMetric;
 
 /**
- * Probability of Connectivity Circuit version.
- * The shortest path is replaced by the resistance of the circuit.
+ * Interaction Flux local metric.
  * 
  * @author Gilles Vuidel
  */
-public class PCCircMetric extends GlobalMetric implements PreCalcMetric<Node> {
+public final class IFLocalMetric extends LocalSingleMetric {
 
     private AlphaParamMetric alphaParam = new AlphaParamMetric();
-    private double metric;
-    private Circuit circuit;
     
     @Override
-    public Double calcPartMetric(Node n1, GraphGenerator g) {
+    public double calcSingleMetric(Graphable g, GraphGenerator gen) {
+        Node node = (Node) g;
+        double srcCapa = Project.getPatchCapacity(node);
+        GraphPathFinder pathFinder = gen.getPathFinder(node);
         double sum = 0;
-        double srcCapa = Project.getPatchCapacity(n1);
-        for(Node n2 : g.getNodes()) {
-            double r = circuit.computeR(n1, n2);
-            sum += Math.pow(srcCapa * Project.getPatchCapacity(n2), alphaParam.getBeta()) * Math.exp(-alphaParam.getAlpha()*r);
-        }
+        for(Node n : pathFinder.getComputedNodes()) {
+            sum += Math.exp(-alphaParam.getAlpha() * pathFinder.getCost(n)) * Math.pow(srcCapa * Project.getPatchCapacity(n), alphaParam.getBeta());            
+        }            
+        
         return sum;
     }
 
     @Override
-    public Double[] calcMetric(GraphGenerator g) {
-        return new Double[]{metric};
-    }
-
-    @Override
-    public void startCalc(GraphGenerator g) {
-        metric = 0;
-        circuit = new Circuit(g);
-    }
-
-    @Override
-    public TypeParam getTypeParam() {
-        return TypeParam.NODE;
-    }
-    
-    @Override
-    public void mergePart(Object part) {
-        metric += (Double)part;
-    }
-
-    @Override
-    public void endCalc(GraphGenerator g) {
-        circuit = null;
-        metric = metric / Math.pow(g.getProject().getArea(), 2);       
-    }
-    
-    @Override
     public String getShortName() {
-        return "PCCirc";
+        return "IF";
+    }
+
+    @Override
+    public boolean calcNodes() {
+        return true;
     }
     
     @Override
@@ -104,6 +81,4 @@ public class PCCircMetric extends GlobalMetric implements PreCalcMetric<Node> {
     public Type getType() {
         return Type.WEIGHT;
     }
-
-
 }